+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## 3.8.0 (2025-11-12)
 
 - Add new cop `RSpec/LeakyLocalVariable`. ([@lovro-bikic])
+- Add new cop `RSpec/RedundantPending`. ([@ydah])
 - Bump RuboCop requirement to +1.81. ([@ydah])
 - Fix a false positive for `RSpec/LetSetup` when `let!` used in outer scope. ([@ydah])
 - Fix a false positive for `RSpec/ReceiveNever` cop when `allow(...).to receive(...).never`. ([@ydah])